refactor: 변경된 로고 적용 - 서랍장 헤더 - 로그인 - 검색바
